remove fromIntegral

+++ src/GrowthOfPopulation.hs
@@ -1,7 +1,9 @@
 module GrowthOfPopulation where
 
+-- https://www.codewars.com/kata/563b662a59afc2b5120000c6
+
 nbYear :: Int -> Double -> Int -> Int -> Int
-nbYear p0 percent aug p = go 0 (fromIntegral p0)
+nbYear p0 percent aug p = go 0 p0
   where
     go :: Int -> Int -> Int
     go acc x
